+++ app/Http/Controllers/Api/PermissionsController.php
@@ -15,6 +15,7 @@ class PermissionsController extends Controller
      */
     public function index()
     {
+        abort_if(Gate::denies('permission_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         $permissions = Permission::orderBy('id', 'desc')->paginate(8);
         return response()->json($permissions);
     }
@@ -27,6 +28,7 @@ class PermissionsController extends Controller
      */
     public function store(Request $request)
     {
+        abort_if(Gate::denies('permission_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         $permissions = Permission::create($request->all());
         // $customers = json_encode($request->customers);
         // $members->customers()->sync(json_decode($customers));
@@ -45,6 +47,7 @@ class PermissionsController extends Controller
      */
     public function show(Permission $permission)
     {
+        abort_if(Gate::denies('permission_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
         return response()->json($permission);
     }
@@ -58,6 +61,7 @@ class PermissionsController extends Controller
      */
     public function update(Request $request, Permission $permission)
     {
+        abort_if(Gate::denies('permission_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         $input = $request->all();
         $permission->update($input);
         // $customers = json_encode($request->customers);
@@ -78,6 +82,7 @@ class PermissionsController extends Controller
      */
     public function destroy($id)
     {
+        abort_if(Gate::denies('permission_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         $permission = Permission::find($id);
 
         $permission->destroy($id);
